Some small changes

+++ sssom/writers.py
@@ -50,7 +50,7 @@ def write_tsv(
     msdoc = to_mapping_set_document(msdf)
     df = to_dataframe(msdf)
     meta = extract_global_metadata(msdoc)
-    if(filename):
+    if filename and filename != "-":
         if os.path.isfile(filename):
             os.remove(filename)
         f = open(filename, "a")
@@ -62,8 +62,11 @@ def write_tsv(
         f.close()
     else:
         # stdout the result for now
-        sys.stdout.write(str(meta))
-        sys.stdout.write(str(df))
+        if meta:
+            mapping_data_string = yaml.dump(meta)
+            for line in mapping_data_string.splitlines():
+                sys.stdout.write("#" + line + "\n")
+        df.to_csv(sys.stdout, sep=sep, index=False)
 
 
 def write_rdf(
